# contracts/salt-control-api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Ring0ApprovalRequest:
      properties:
        decision:
          default: approve
          title: Decision
          type: string
        reason:
          default: ''
          title: Reason
          type: string
        role:
          title: Role
          type: string
      required:
      - role
      title: Ring0ApprovalRequest
      type: object
    Ring0CreateRequest:
      properties:
        component:
          title: Component
          type: string
        configRevision:
          title: Configrevision
          type: string
        releaseId:
          title: Releaseid
          type: string
        requestId:
          title: Requestid
          type: string
        targets:
          items:
            additionalProperties: true
            type: object
          title: Targets
          type: array
        version:
          title: Version
          type: string
      required:
      - component
      - version
      - requestId
      - releaseId
      - configRevision
      title: Ring0CreateRequest
      type: object
    Ring0RollbackRequest:
      properties:
        endpointId:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpointid
        reason:
          default: ''
          title: Reason
          type: string
        scope:
          default: target
          title: Scope
          type: string
      title: Ring0RollbackRequest
      type: object
...
    RolloutActionRequest:
      properties:
        reason:
          title: Reason
          type: string
        requestId:
          title: Requestid
          type: string
      required:
      - requestId
      - reason
      title: RolloutActionRequest
      type: object
    RolloutApprovalRequest:
      properties:
        decision:
          default: approve
          title: Decision
          type: string
        reason:
          default: ''
          title: Reason
          type: string
        requestId:
          title: Requestid
          type: string
      required:
      - requestId
      title: RolloutApprovalRequest
      type: object
    RolloutCreateRequest:
      properties:
        component:
          title: Component
          type: string
        filters:
          additionalProperties: true
          default: {}
          title: Filters
          type: object
        reason:
          default: ''
          title: Reason
          type: string
        requestId:
          title: Requestid
          type: string
        ring:
          title: Ring
          type: string
        thresholds:
          additionalProperties: true
          default: {}
          title: Thresholds
          type: object
        version:
          title: Version
          type: string
      required:
      - component
      - version
      - ring
      - requestId
      title: RolloutCreateRequest
      type: object
    RolloutResponse:
      properties:
        approvalRequired:
          default: true
          title: Approvalrequired
          type: boolean
        batchSize:
          default: 0
          title: Batchsize
          type: integer
        completedCount:
          default: 0
          title: Completedcount
          type: integer
        component:
          title: Component
          type: string
        failureRate:
          default: 0.0
          title: Failurerate
          type: number
        p0Count:
          default: 0
          title: P0Count
          type: integer
        p1Count:
          default: 0
          title: P1Count
          type: integer
        ring:
          title: Ring
          type: string
        rollbackRate:
          default: 0.0
          title: Rollbackrate
          type: number
        rolloutId:
          title: Rolloutid
          type: string
        state:
          title: State
          type: string
        successRate:
          default: 0.0
          title: Successrate
          type: number
        targetCount:
          default: 0
          title: Targetcount
          type: integer
        thresholds:
          additionalProperties: true
          default: {}
          title: Thresholds
          type: object
        version:
          title: Version
          type: string
      required:
      - rolloutId
      - component
      - version
      - ring
      - state
      title: RolloutResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: SMC Salt Control
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /salt/v1/ring0/rollouts:
    post:
      operationId: create_ring0_salt_v1_ring0_rollouts_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ring0CreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Ring0 Salt V1 Ring0 Rollouts Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Ring0
      tags:
      - ring0
...
  /salt/v1/ring0/rollouts/{rollout_id}:approve:
    post:
      operationId: approve_ring0_salt_v1_ring0_rollouts__rollout_id__approve_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ring0ApprovalRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Approve Ring0 Salt V1 Ring0 Rollouts  Rollout Id  Approve Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Approve Ring0
      tags:
      - ring0
  /salt/v1/ring0/rollouts/{rollout_id}:rollback:
    post:
      operationId: rollback_ring0_salt_v1_ring0_rollouts__rollout_id__rollback_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ring0RollbackRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Rollback Ring0 Salt V1 Ring0 Rollouts  Rollout Id  Rollback Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rollback Ring0
      tags:
      - ring0
...
  /salt/v1/rollouts:
    post:
      operationId: create_rollout_salt_v1_rollouts_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:
    get:
      operationId: get_rollout_salt_v1_rollouts__rollout_id__get
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:abort:
    post:
      operationId: abort_rollout_salt_v1_rollouts__rollout_id__abort_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Abort Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:advance:
    post:
      operationId: advance_rollout_salt_v1_rollouts__rollout_id__advance_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Advance Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:approve:
    post:
      operationId: approve_rollout_salt_v1_rollouts__rollout_id__approve_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutApprovalRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Approve Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:pause:
    post:
      operationId: pause_rollout_salt_v1_rollouts__rollout_id__pause_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Pause Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:resume:
    post:
      operationId: resume_rollout_salt_v1_rollouts__rollout_id__resume_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resume Rollout
      tags:
      - rollouts
  /salt/v1/rollouts/{rollout_id}:rollback:
    post:
      operationId: rollback_rollout_salt_v1_rollouts__rollout_id__rollback_post
      parameters:
      - in: path
        name: rollout_id
        required: true
        schema:
          title: Rollout Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutActionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rollback Rollout
      tags:
      - rollouts
...